+++ trifinger_simulation/tasks/move_cube/evaluate_policy.py
@@ -24,7 +24,7 @@ evaluate the actual performance of the policy.
 """
 import sys
 
-import gym
+import gymnasium as gym
 
 from trifinger_simulation.gym_wrapper.envs import cube_env
 from trifinger_simulation.tasks import move_cube
@@ -81,11 +81,12 @@ def main():
     # environment, this is the case when looping until is_done == True.  Make
     # sure to adjust this in case your custom environment behaves differently!
     is_done = False
-    observation = env.reset()
+    observation, info = env.reset()
     accumulated_reward = 0
     while not is_done:
         action = policy.predict(observation)
-        observation, reward, is_done, info = env.step(action)
+        observation, reward, terminated, truncated, info = env.step(action)
+        is_done = terminated or truncated
         accumulated_reward += reward
 
     print("Accumulated reward: {}".format(accumulated_reward))